refactor: remove duplicated hook
